// packages/server/src/types.ts
export interface WebSocketLike {
	send(data: string | ArrayBufferLike | Blob | ArrayBufferView): void;
	close(code?: number, reason?: string): void;
	readyState: number;
}
// ...
export interface Extension<Context = any> {
	priority?: number;
	extensionName?: string;
	onConfigure?(data: onConfigurePayload): Promise<any>;
	onListen?(data: onListenPayload): Promise<any>;
	onUpgrade?(data: onUpgradePayload): Promise<any>;
	onConnect?(data: onConnectPayload<Context>): Promise<any>;
	connected?(data: connectedPayload<Context>): Promise<any>;
	onAuthenticate?(data: onAuthenticatePayload<Context>): Promise<any>;
	onTokenSync?(data: onTokenSyncPayload<Context>): Promise<any>;
	onCreateDocument?(data: onCreateDocumentPayload<Context>): Promise<any>;
	onLoadDocument?(data: onLoadDocumentPayload<Context>): Promise<any>;
	afterLoadDocument?(data: afterLoadDocumentPayload<Context>): Promise<any>;
	beforeHandleMessage?(data: beforeHandleMessagePayload<Context>): Promise<any>;
	beforeSync?(data: beforeSyncPayload<Context>): Promise<any>;
	beforeBroadcastStateless?(
		data: beforeBroadcastStatelessPayload,
	): Promise<any>;
	onStateless?(payload: onStatelessPayload): Promise<any>;
	onChange?(data: onChangePayload<Context>): Promise<any>;
	onStoreDocument?(data: onStoreDocumentPayload<Context>): Promise<any>;
	afterStoreDocument?(data: afterStoreDocumentPayload<Context>): Promise<any>;
	onAwarenessUpdate?(data: onAwarenessUpdatePayload<Context>): Promise<any>;
	onRequest?(data: onRequestPayload): Promise<any>;
	onDisconnect?(data: onDisconnectPayload<Context>): Promise<any>;
	beforeUnloadDocument?(data: beforeUnloadDocumentPayload): Promise<any>;
	afterUnloadDocument?(data: afterUnloadDocumentPayload): Promise<any>;
	onDestroy?(data: onDestroyPayload): Promise<any>;
}
// ...
export interface Configuration<Context = any> extends Extension<Context> {
	/**
	 * A name for the instance, used for logging.
	 */
	name: string | null;
	/**
	 * A list of collabApi extensions.
	 */
	extensions: Array<Extension>;
	/**
	 * Defines in which interval the server sends a ping, and closes the connection when no pong is sent back.
	 */
	timeout: number;
	/**
	 * Debounces the call of the `onStoreDocument` hook for the given amount of time in ms.
	 * Otherwise every single update would be persisted.
	 */
	debounce: number;
	/**
	 * Makes sure to call `onStoreDocument` at least in the given amount of time (ms).
	 */
	maxDebounce: number;
	/**
	 * By default, the servers show a start screen. If passed false, the server will start quietly.
	 */
	quiet: boolean;
	/**
	 * If set to false, respects the debounce time of `onStoreDocument` before unloading a document.
	 * Otherwise, the document will be unloaded immediately.
	 *
	 * This prevents a client from DOSing the server by repeatedly connecting and disconnecting when
	 * your onStoreDocument is rate-limited.
	 */
	unloadImmediately: boolean;

	/**
	 * options to pass to the ydoc document
	 */
	yDocOptions: {
		gc: boolean; // enable or disable garbage collection (see https://github.com/yjs/yjs/blob/main/INTERNALS.md#deletions)
		gcFilter: () => boolean; // will be called before garbage collecting ; return false to keep it
	};
}

export interface onStatelessPayload {
	connection: Connection;
	documentName: string;
	document: Document;
	payload: string;
}

export interface onAuthenticatePayload<Context = any> {
	context: Context;
	documentName: string;
	instance: CollabApi;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	request: Request;
	socketId: string;
	token: string;
	connectionConfig: ConnectionConfiguration;
	providerVersion: string | null;
}

export interface onTokenSyncPayload<Context = any> {
	context: Context;
	document: Document;
	documentName: string;
	instance: CollabApi;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	socketId: string;
	token: string;
	connectionConfig: ConnectionConfiguration;
	connection: Connection<Context>;
}

export interface onCreateDocumentPayload<Context = any> {
	context: Context;
	documentName: string;
	instance: CollabApi;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	socketId: string;
	connectionConfig: ConnectionConfiguration;
}

export interface onConnectPayload<Context = any> {
	context: Context;
	documentName: string;
	instance: CollabApi;
	request: Request;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	socketId: string;
	connectionConfig: ConnectionConfiguration;
	providerVersion: string | null;
}

export interface connectedPayload<Context = any> {
	context: Context;
	documentName: string;
	instance: CollabApi;
	request: Request;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	socketId: string;
	connectionConfig: ConnectionConfiguration;
	connection: Connection<Context>;
	providerVersion: string | null;
}

export interface onLoadDocumentPayload<Context = any> {
	context: Context;
	document: Document;
	documentName: string;
	instance: CollabApi;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	socketId: string;
	connectionConfig: ConnectionConfiguration;
}

export interface afterLoadDocumentPayload<Context = any> {
	context: Context;
	document: Document;
	documentName: string;
	instance: CollabApi;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	socketId: string;
	connectionConfig: ConnectionConfiguration;
}

export interface onChangePayload<Context = any> {
	clientsCount: number;
	context: Context;
	document: Document;
	documentName: string;
	instance: CollabApi;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	update: Uint8Array;
	socketId: string;
	transactionOrigin: unknown;
	connection?: Connection<Context>;
}

export interface beforeHandleMessagePayload<Context = any> {
	clientsCount: number;
	context: Context;
	document: Document;
	documentName: string;
	instance: CollabApi;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	update: Uint8Array;
	socketId: string;
	connection: Connection<Context>;
}

export interface beforeSyncPayload<Context = any> {
	clientsCount: number;
	context: Context;
	document: Document;
	documentName: string;
	connection: Connection<Context>;
	/**
	 * The y-protocols/sync message type
	 * @example
	 * 0: SyncStep1
	 * 1: SyncStep2
	 * 2: YjsUpdate
	 *
	 * @see https://github.com/yjs/y-protocols/blob/master/sync.js#L13-L40
	 */
	type: number;
	/**
	 * The payload of the y-sync message.
	 */
	payload: Uint8Array;
}

export interface beforeBroadcastStatelessPayload {
	document: Document;
	documentName: string;
	payload: string;
}

export interface onStoreDocumentPayload<Context = any> {
	clientsCount: number;
	document: Document;
	lastContext: Context;
	lastTransactionOrigin: unknown;
	documentName: string;
	instance: CollabApi;
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface, @typescript-eslint/no-empty-object-type
export interface afterStoreDocumentPayload<Context = any>
	extends onStoreDocumentPayload<Context> {}

export interface onAwarenessUpdatePayload<Context = any> {
	document: Document;
	documentName: string;
	instance: CollabApi;
	transactionOrigin: unknown;
	connection?: Connection<Context>;
	added: number[];
	updated: number[];
	removed: number[];
	awareness: Awareness;
	states: StatesArray;
}
// ...
export interface onDisconnectPayload<Context = any> {
	clientsCount: number;
	context: Context;
	document: Document;
	documentName: string;
	instance: CollabApi;
	requestHeaders: Headers;
	requestParameters: URLSearchParams;
	socketId: string;
}

export interface onRequestPayload {
	request: IncomingMessage;
	response: ServerResponse;
	instance: CollabApi;
}

export interface onUpgradePayload {
	request: IncomingMessage;
	socket: any;
	head: any;
	instance: CollabApi;
}

export interface onListenPayload {
	instance: CollabApi;
	configuration: Configuration;
	port: number;
}

export interface onDestroyPayload {
	instance: CollabApi;
}

export interface onConfigurePayload {
	instance: CollabApi;
	configuration: Configuration;
	version: string;
}

export interface afterUnloadDocumentPayload {
	instance: CollabApi;
	documentName: string;
}

export interface beforeUnloadDocumentPayload {
	instance: CollabApi;
	documentName: string;
	document: Document;
}